backend/pyproject.toml
[project]
name = "quang-long-api"
version = "0.1.0"
description = "Ghe Da Quang Long API"
authors = [
    {name = "Quang Long", email = "ghedaquanglong@gmail.com"}
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.24.0",
    "motor>=3.3.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "passlib[bcrypt]",
    "python-jose[cryptography]",
    "bcrypt>=4.0.1",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "B", "I"]

[tool.ruff.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=app"

[tool.coverage.run]
source = ["app"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "pass",
    "@abstractmethod",
]
